asp-validation nas telas necessárias

+++ Models/Nota.cs
@@ -6,10 +6,10 @@ public class Nota
     public int? CdNota{get;set;}
     public int? CdAluno{get;set;}
     public int? CdTurmaProfessor{get;set;}
-    public int Nota1 {get;set;}
-    public int Nota2 {get;set;}
-    public int Nota3 {get;set;}
-    public int Nota4 {get;set;}
+    public float Nota1 {get;set;}
+    public float Nota2 {get;set;}
+    public float Nota3 {get;set;}
+    public float Nota4 {get;set;}
 }
 
 public class NotaSubmit : Nota
